requirements.txt
langchain-google-genai>=2.1.0 
python-dotenv>=1.0.0 
langgraph>=0.5.0
langgraph-checkpoint-sqlite>=2.0.0
langchain-community>=0.3.0
langchain_chroma>=0.1.0
sentence-transformers>=2.2.0
fastapi>=0.116.1
python-multipart>=0.0.6
openpyxl>=3.1.2
xlrd>=2.0.1
pandas>=2.2.2